Update mailer_patch.rb
After this modification Redmine notification mails seen as from user. But, reply adres does not change.
This is more usable for us. I don't know the ruby. My code is specified for us. 
"/Portal/" is suffix of the user name. it is permit some error when user name have non-english characters. secondly, It gives a message to recipients about source of email.

Sorry for my english.